fix: some fixes after review
