+++ playwright-tests/.env.example
@@ -4,7 +4,7 @@
 STAND_ENV=testnet
 
 # Link type for a test run. Can be equal: stand, localhost
-STAND_TYPE=stand
+STAND_TYPE=localhost
 
 # Wallet configurations
 ## Secret phase of the wallet to set up the wallet extension
